requirements.txt
django
South
django-grappelli
django-debug-toolbar
django-registration
django-blog-zinnia
Pillow